Add an optional :offload send mode
This optional mode offloads the send/2 calls to the remote Manifold nodes
to one of the local `Manifold.Partitioner` processes.

Those processes on the send side are otherwise unused. This gives the
caller an option to further free up the sender process.

Add a new test for the :offload send mode.